refactor(clients): extract low-fanout config export formats
